#18 Move /core/security.py to /services/auth/security_service.py as it is more of a service rather than app core configuration
