+++ backend/src/database/crud/user_crud.py
@@ -1,4 +1,4 @@
-from backend.src.core.security import hash_password
+from backend.src.services.auth.security_service import hash_password
 from sqlalchemy.orm.session import Session
 from backend.src.database.models.user_model import User
 from backend.src.schemas.model.user_schema import UserCreate
